CNS-382: fix double prefix of entry-index
The entry-index was mistakenly stored with double prefix, so fix it. (For
example, for some <index>, change "Entry_Index_prj-fsEntry_Index_prj-fs<index>"
to be "Entry_Index_prj-fs<index>".

Also rename some functions for better code clarity.
